Add Checker class to scan for refurbished match
This commit adds a simple checker for scanning Apple Canada's
refurbished macbook airs for a matching item. If it's present and the
price is under a specified value, it will send an SMS via Twilio to
notify you.

Scheduler
====================
I've deployed this via Heroku and made use of the Heroku Scheduler to
run `ruby run_checker.rb` every hour.